requirements.txt
# ---------------------------
# Web app
# ---------------------------
streamlit>=1.29.0

# ---------------------------
# Data handling
# ---------------------------
pandas>=2.1.1
numpy>=1.26.0

# ---------------------------
# PDF & DOCX parsing
# ---------------------------
PyPDF2>=3.0.1
python-docx>=1.1.0
PyMuPDF>=1.22.5

# ---------------------------
# Visualization
# ---------------------------
plotly>=5.16.0
matplotlib>=3.8.0
seaborn>=0.13.2

# ---------------------------
# Machine learning & NLP
# ---------------------------
scikit-learn>=1.3.2
joblib>=1.3.2
nltk>=3.9.1
spacy>=3.6.1
transformers>=4.42.2
torch>=2.3.0  # PyTorch CPU version
